updates: eslint script
