+++ apps/nextjs/package.json
@@ -28,17 +28,14 @@
   },
   "devDependencies": {
     "@greed/tsconfig": "workspace:*",
+    "eslint-config-custom": "workspace:*",
     "@prisma/nextjs-monorepo-workaround-plugin": "^4.13.0",
-    "@types/eslint": "^8.21.3",
     "@types/node": "^18.15.5",
     "@types/prettier": "^2.7.2",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.56.0",
-    "@typescript-eslint/parser": "^5.56.0",
     "autoprefixer": "^10.4.14",
     "eslint": "^8.36.0",
-    "eslint-config-next": "^13.2.4",
     "postcss": "^8.4.21",
     "prettier": "^2.8.6",
     "prettier-plugin-tailwindcss": "^0.2.6",
