+++ apps/nextjs/src/pages/api/docs.ts
@@ -2,7 +2,7 @@ import type { NextApiRequest, NextApiResponse } from "next";
 import { renderTrpcPanel } from "trpc-panel";
 import { appRouter } from "../../server/api/root";
 
-export default async function handler(_: NextApiRequest, res: NextApiResponse) {
+export default function handler(_: NextApiRequest, res: NextApiResponse) {
   res.status(200).send(
     renderTrpcPanel(appRouter, {
       url: "http://localhost:4001/api/trpc",
