+++ packages/db/package.json
@@ -17,6 +17,7 @@
     "zod": "^3.21.4"
   },
   "devDependencies": {
+    "eslint": "^7.32.0",
     "@greed/tsconfig": "workspace:*",
     "prisma": "^4.11.0",
     "prisma-dbml-generator": "^0.10.0",
