+++ packages/eslint-config-custom/package.json
@@ -10,10 +10,12 @@
     "eslint-config-next": "latest",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-react": "7.28.0",
-    "eslint-config-turbo": "latest"
+    "eslint-config-turbo": "latest",
+    "@typescript-eslint/eslint-plugin": "^5.56.0",
+    "@typescript-eslint/parser": "^5.56.0"
   },
   "devDependencies": {
-    "typescript": "^5.0.2"
+    "@greed/tsconfig": "workspace:*"
   },
   "publishConfig": {
     "access": "public"
