+++ packages/permissions/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "clean": "rm -rf .turbo node_modules",
     "permissions:testing": "vitest --ui",
-    "test": "vitest"
+    "test": "vitest run"
   },
   "dependencies": {
     "zod": "^3.21.4"
